+++ unit-tests/conftest.py
@@ -10,13 +10,15 @@ It manages:
 - Device selection based on markers
 - Context filtering to ensure tests only see intended devices
 - Session-scoped device management
+
+Implementation is split across rspy.pytest sub-modules; this file keeps only the hooks
+and fixtures that pytest requires in conftest.py for auto-discovery.
 """
 
 import pytest
 import sys
 import os
-import re
-from typing import List
+import logging
 
 # unit-tests/py/ contains rspy — the shared helper library used by all RealSense tests
 current_dir = os.path.dirname(os.path.abspath(__file__))
@@ -27,70 +29,23 @@ if py_dir not in sys.path:
 # Consume --debug before any rspy imports (rspy.log also consumes it from sys.argv)
 _debug_requested = '--debug' in sys.argv
 
-import logging
-from rspy import devices, repo, log as rspy_log
+from rspy import devices, repo
 from rspy.signals import register_signal_handlers
+from rspy.pytest.logging_setup import (
+    setup_test_logging, bridge_rspy_log, ensure_newline, configure_logging,
+    start_test_log, stop_test_log, print_terminal_summary,
+)
+from rspy.pytest.cli import consume_legacy_flags
+from rspy.pytest.device_helpers import find_matching_devices, resolve_device_each_serials
+from rspy.pytest.collection import filter_and_sort_items
 
 log = logging.getLogger('librealsense')
 
+# Bridge rspy.log → Python logging early, before any test output
+bridge_rspy_log()
 
-# ============================================================================
-# Bridge rspy.log → Python logging
-# ============================================================================
-# rspy.log writes directly to stdout via its own out() function, so hub messages
-# (port enable/disable, device added/removed) don't appear in per-test log files.
-# Wrap each rspy.log level function to also forward to Python logging.
-
-def _bridge_rspy_log():
-    """Wrap rspy.log.d/i/w/e to also emit via Python logging."""
-    _rspy_module = rspy_log
-
-    def _wrap(original_fn, py_level):
-        def wrapper(*args):
-            result = original_fn(*args)
-            msg = ' '.join(str(a) for a in args)
-            log.log(py_level, msg)
-            return result
-        return wrapper
-
-    # d() is dynamically redefined by debug_on(), so wrap whatever is current
-    _rspy_module.d = _wrap(_rspy_module.d, logging.DEBUG)
-    _rspy_module.i = _wrap(_rspy_module.i, logging.INFO)
-    _rspy_module.w = _wrap(_rspy_module.w, logging.WARNING)
-    _rspy_module.e = _wrap(_rspy_module.e, logging.ERROR)
-
-_bridge_rspy_log()
-
-
-# ============================================================================
-# Legacy CLI Flag Translation
-# ============================================================================
-# The old run-unit-tests.py used flags like -r/--regex that clash with pytest built-ins.
-# We intercept and translate them here, before pytest parses sys.argv.
-
-def _find_flag(flag):
-    """Find a flag in sys.argv, returning its index or None."""
-    try:
-        return sys.argv.index(flag)
-    except ValueError:
-        return None
-
-
-def _consume_flag_with_arg(flags, pytest_equiv):
-    """Consume a flag+argument from sys.argv, translate to pytest equivalent."""
-    for flag in flags:
-        idx = _find_flag(flag)
-        if idx is not None:
-            if idx + 1 >= len(sys.argv):
-                print(f'-F- {flag} requires an argument', file=sys.stderr)
-                sys.exit(1)
-            value = sys.argv[idx + 1]
-            del sys.argv[idx:idx + 2]
-            sys.argv.extend([pytest_equiv, value])
-            return value
-    return None
-
-_consume_flag_with_arg(['-r', '--regex'], '-k')  # -r/--regex -> pytest's -k (keyword filter)
+# Translate legacy CLI flags before pytest parses sys.argv
+consume_legacy_flags()
 
 
 # ============================================================================
@@ -106,82 +61,6 @@ try:
 except ImportError:
     log.warning('No pyrealsense2 library available!')
     rs = None
-
-
-# ============================================================================
-# Platform Detection
-# ============================================================================
-
-def is_jetson_platform():
-    """Detect NVIDIA Jetson — some tests behave differently on embedded platforms."""
-    try:
-        with open('/proc/device-tree/model', 'r') as f:
-            model = f.read()
-            return 'jetson' in model.lower()
-    except:
-        return False
-
-
-# ============================================================================
-# Logging Setup
-# ============================================================================
-
-def _find_build_dir():
-    """Walk up from unit-tests/ to find the CMake build dir (contains CMakeCache.txt)."""
-    search_dir = current_dir
-    while True:
-        cmake_cache = os.path.join(search_dir, 'CMakeCache.txt')
-        if os.path.isfile(cmake_cache):
-            log.debug(f'Found build dir: {search_dir}')
-            return search_dir
-        parent = os.path.dirname(search_dir)
-        if parent == search_dir:
-            if hasattr(repo, 'build') and repo.build:
-                log.debug(f'Using repo.build: {repo.build}')
-                return repo.build
-            break
-        search_dir = parent
-
-    log.debug('Could not find build directory, using default')
-    return None
-
-
-def _setup_test_logging(config):
-    """Set up per-test log directory and JUnit XML output path (<build_dir>/<config>/unit-tests/)."""
-    build_dir = _find_build_dir()
-
-    if build_dir:
-        cmake_cache_path = os.path.join(build_dir, 'CMakeCache.txt')
-        configuration = None
-
-        try:
-            with open(cmake_cache_path, 'r') as f:
-                for line in f:
-                    if line.startswith('CMAKE_BUILD_TYPE:'):
-                        parts = line.split('=', 1)
-                        if len(parts) == 2:
-                            configuration = parts[1].strip()
-                            log.debug(f'Found CMAKE_BUILD_TYPE: {configuration}')
-                            break
-        except Exception as e:
-            log.debug(f'Could not read CMAKE_BUILD_TYPE from CMakeCache.txt: {e}')
-
-        if configuration:
-            logdir = os.path.join(build_dir, configuration, 'unit-tests')
-        else:
-            logdir = os.path.join(build_dir, 'unit-tests')
-    else:
-        logdir = os.path.join(current_dir, 'logs')
-
-    os.makedirs(logdir, exist_ok=True)
-    log.debug(f'Test logs directory: {logdir}')
-
-    if not config.getoption('--junitxml', default=None):
-        junit_xml_path = os.path.join(logdir, 'pytest-results.xml')
-        config.option.xmlpath = junit_xml_path
-        log.info(f'JUnit XML results: {junit_xml_path}')
-
-    config._test_logdir = logdir
 
 
 # ============================================================================
@@ -261,7 +140,7 @@ def pytest_configure(config):
         log.info(f"Test context: {context_list}")
 
     # Set up test log directory
-    _setup_test_logging(config)
+    setup_test_logging(config)
 
     # Test discovery defaults (replaces pytest.ini which is .gitignored)
     config.addinivalue_line("python_files", "pytest-*.py")
@@ -315,22 +194,7 @@ def pytest_configure(config):
     )
 
     # Configure standard logging with format matching legacy rspy.log output
-    global _debug_requested
-    if not _debug_requested:
-        log_cli_level = config.getoption('--log-cli-level', default=None)
-        if log_cli_level and log_cli_level.upper() == 'DEBUG':
-            _debug_requested = True
-    log_level_name = 'DEBUG' if _debug_requested else 'INFO'
-    logging.getLogger().setLevel(getattr(logging, log_level_name))
-    capture = config.getoption('capture', default='fd')
-    global _live_logging
-    if capture == 'no':  # -s passed: stream logs to console
-        _live_logging = True
-        config.option.log_cli_level = log_level_name
-        config.option.log_cli_format = '-%(levelname).1s- %(message)s'
-        config.option.log_cli_date_format = ''
-    if _debug_requested:
-        logging.getLogger('paramiko').setLevel(logging.WARNING)
+    configure_logging(config, _debug_requested)
 
     # Query devices early for test parametrization
     try:
@@ -343,153 +207,28 @@ def pytest_configure(config):
 
 
 def pytest_generate_tests(metafunc):
-    """Expand @device_each into one test instance per matching device (e.g. test[D455-SN123])."""
-    device_each_markers = [m for m in metafunc.definition.iter_markers("device_each")]
-
-    if not device_each_markers:
-        return
-
-    all_serials = []
-
-    # Resolve exclusion patterns (markers + CLI) to a set of excluded serial numbers
-    exclude_markers = [m for m in metafunc.definition.iter_markers("device_exclude")]
-    exclude_patterns = [m.args[0] for m in exclude_markers if m.args]
-    cli_excludes = metafunc.config.getoption("--device-exclude", default=[])
-    exclude_patterns.extend(cli_excludes)
-    excluded_sns = set()
-    for pattern in exclude_patterns:
-        excluded_sns.update(devices.by_spec(pattern, []))
-
-    # Resolve CLI --device includes to a set of allowed serial numbers (None = no filter)
-    cli_includes = metafunc.config.getoption("--device", default=[])
-    included_sns = None
-    if cli_includes:
-        included_sns = set()
-        for inc in cli_includes:
-            included_sns.update(devices.by_spec(inc, []))
-
-    for marker in device_each_markers:
-        if not marker.args:
-            continue
-        pattern = marker.args[0]
-        for sn in devices.by_spec(pattern, []):
-            if sn in excluded_sns:
-                continue
-            if included_sns is not None and sn not in included_sns:
-                continue
-            if sn not in all_serials:
-                all_serials.append(sn)
-
-    if all_serials:
-        ids = [f"{devices.get(sn).name}-{sn}" for sn in all_serials]
-        metafunc.fixturenames.append('_test_device_serial')
-        metafunc.parametrize("_test_device_serial", all_serials, ids=ids, scope="function")
+    """Expand @device_each into one test instance per matching device."""
+    resolve_device_each_serials(metafunc)
 
 
 def pytest_collection_modifyitems(config, items):
-    """Auto-skip nightly/dds tests unless opted in, filter --live, and sort by priority."""
-    markexpr = config.getoption("-m", default="")
-
-    if not (markexpr and "nightly" in markexpr):
-        skip_nightly = pytest.mark.skip(reason="Nightly test (use -m nightly to run)")
-        for item in items:
-            if "nightly" in item.keywords:
-                item.add_marker(skip_nightly)
-
-    if not (markexpr and "dds" in markexpr):
-        skip_dds = pytest.mark.skip(reason="DDS test (use -m dds to run)")
-        for item in items:
-            if "dds" in item.keywords:
-                item.add_marker(skip_dds)
-
-    # Skip non-device tests when --live is specified
-    if config.getoption("--live", default=False):
-        skip_no_device = pytest.mark.skip(reason="--live: test has no device requirement")
-        for item in items:
-            has_device = any(item.iter_markers("device")) or any(item.iter_markers("device_each"))
-            if not has_device:
-                item.add_marker(skip_no_device)
-
-    def get_priority(item):
-        marker = item.get_closest_marker("priority")
-        if marker and marker.args:
-            return marker.args[0]
-        return 500
-
-    items.sort(key=get_priority)
-
-    # Group parametrized tests by device within each module, so all tests run on one
-    # device before switching to the next (matching run-unit-tests.py behavior).
-    def get_device_group_key(item):
-        module = item.module.__name__
-        if hasattr(item, 'callspec') and '_test_device_serial' in item.callspec.params:
-            device_serial = item.callspec.params['_test_device_serial']
-        else:
-            device_serial = ''
-        return (module, device_serial)
-
-    items.sort(key=get_device_group_key)
-
-
-_live_logging = False  # Set in pytest_configure when -s is passed
-
-def _ensure_newline():
-    """Pytest's progress dots (F/.) don't end with newline — force one before our log output.
-    Only needed when live logging is active (-s), otherwise it breaks -v formatting."""
-    if _live_logging:
-        sys.stdout.write('\n')
-        sys.stdout.flush()
-
-
-def _test_log_name(item):
-    """Convert a node id like 'live/frames/pytest-t2ff.py::test_x[D455-SN]' to a log filename."""
-    # Get the test file basename without extension
-    file_path = item.fspath
-    basename = os.path.splitext(os.path.basename(str(file_path)))[0]  # 'pytest-t2ff-pipeline'
-
-    # Get the test name with parametrize suffix
-    test_name = item.name  # 'test_func[D455-SN]'
-
-    # Combine: pytest-t2ff-pipeline_test_func[D455-SN].log
-    # Sanitize for filesystem
-    log_name = f"{basename}_{test_name}"
-    log_name = re.sub(r'[<>:"/\\|?*]', '_', log_name)
-    return log_name + ".log"
+    """Auto-skip nightly/dds tests, filter --live, sort by priority."""
+    filter_and_sort_items(config, items)
 
 
 @pytest.hookimpl(hookwrapper=True)
 def pytest_runtest_protocol(item, nextitem):
-    """Wrap each test with log separators and write per-test log file via logging FileHandler."""
-    _ensure_newline()
+    """Wrap each test with log separators and write per-test log file."""
+    ensure_newline()
     log.info("-" * 80)
     log.info(f"Test: {item.nodeid}")
     log.info("-" * 80)
 
-    # Add per-test log file handler (only when stdout is captured, i.e. -s not passed)
-    logdir = getattr(item.config, '_test_logdir', None)
-    file_handler = None
-    capture = item.config.getoption('capture', default='fd')
-
-    if logdir and capture != 'no':
-        log_name = _test_log_name(item)
-        log_path = os.path.join(logdir, log_name)
-        try:
-            file_handler = logging.FileHandler(log_path, mode='w')
-            file_handler.setFormatter(logging.Formatter('-%(levelname).1s- %(message)s'))
-            file_handler.setLevel(logging.DEBUG)
-            logging.getLogger().addHandler(file_handler)
-        except Exception as e:
-            log.warning(f"Could not create test log file {log_path}: {e}")
-            file_handler = None
-
+    file_handler = start_test_log(item)
     outcome = yield
+    stop_test_log(file_handler)
 
-    # Remove per-test file handler
-    if file_handler is not None:
-        logging.getLogger().removeHandler(file_handler)
-        file_handler.close()
-
-    _ensure_newline()
+    ensure_newline()
 
 
 @pytest.hookimpl(hookwrapper=True)
@@ -499,42 +238,13 @@ def pytest_runtest_makereport(item, call):
     report = outcome.get_result()
 
     if call.when == "call":
-        _ensure_newline()
+        ensure_newline()
         log.debug(f"Test execution took {report.duration:.3f}s")
 
 
 def pytest_terminal_summary(terminalreporter, exitstatus, config):
-    """Print a clear pass/fail/skip summary at the end of the run.
-
-    Uses both log.info (captured in per-test log files) and print() (always on stdout)
-    so Jenkins Groovy can parse the '-I-' prefixed lines from the tee'd console log.
-    """
-    _ensure_newline()
-
-    passed = len(terminalreporter.stats.get('passed', []))
-    failed = len(terminalreporter.stats.get('failed', []))
-    skipped = len(terminalreporter.stats.get('skipped', []))
-    error = len(terminalreporter.stats.get('error', []))
-    total = passed + failed + skipped + error
-
-    def summary(msg):
-        log.info(msg)
-        print(f"-I- {msg}")
-
-    summary("")
-    summary("=" * 80)
-    summary("Test Summary")
-    summary("=" * 80)
-    summary(f"Total tests run: {total}")
-    if passed > 0:
-        summary(f"Passed: {passed}")
-    if failed > 0:
-        summary(f"Failed: {failed}")
-    if skipped > 0:
-        summary(f"Skipped: {skipped}")
-    if error > 0:
-        summary(f"Errors: {error}")
-    summary("=" * 80)
+    """Print pass/fail/skip summary for Jenkins Groovy parsing."""
+    print_terminal_summary(terminalreporter)
 
 
 # ============================================================================
@@ -579,7 +289,7 @@ def session_setup_teardown():
 
     yield
 
-    _ensure_newline()
+    ensure_newline()
     log.info("")
     log.info("=" * 80)
     log.info("Pytest Session Ending")
@@ -624,7 +334,7 @@ def module_device_setup(request):
             yield None
             return
 
-        serial_numbers = _find_matching_devices(device_markers, each=False,
+        serial_numbers = find_matching_devices(device_markers, each=False,
                                                   cli_includes=request.config.getoption("--device", default=[]),
                                                   cli_excludes=request.config.getoption("--device-exclude", default=[]))
 
@@ -693,65 +403,7 @@ def test_device(test_context):
     return dev, test_context
 
 
-
 @pytest.fixture
 def test_context_var():
     """Expose the --context tags (e.g. ['nightly', 'weekly']) so tests can branch on them."""
     return context_list
-
-
-# ============================================================================
-# Helper Functions
-# ============================================================================
-
-def _find_matching_devices(device_markers, each=True, cli_includes=None, cli_excludes=None) -> List[str]:
-    """Resolve device markers + CLI filters into a list of matching serial numbers."""
-    matching_sns = []
-
-    if cli_includes is None:
-        cli_includes = []
-    if cli_excludes is None:
-        cli_excludes = []
-
-    # Resolve exclusion patterns (markers + CLI) to a set of excluded serial numbers
-    exclude_patterns = []
-    for marker in device_markers:
-        if marker.name == 'device_exclude' and marker.args:
-            exclude_patterns.append(marker.args[0])
-            log.debug(f"Excluding devices matching pattern: {marker.args[0]}")
-    exclude_patterns.extend(cli_excludes)
-
-    excluded_sns = set()
-    for pattern in exclude_patterns:
-        excluded_sns.update(devices.by_spec(pattern, []))
-
-    # Resolve CLI includes to a set of allowed serial numbers (None = no filter)
-    included_sns = None
-    if cli_includes:
-        included_sns = set()
-        for inc in cli_includes:
-            included_sns.update(devices.by_spec(inc, []))
-
-    # Find matching devices
-    for marker in device_markers:
-        if marker.name not in ['device', 'device_each'] or not marker.args:
-            continue
-
-        pattern = marker.args[0]
-        log.debug(f"Looking for devices matching pattern: {pattern}")
-
-        for sn in devices.by_spec(pattern, []):
-            if sn in excluded_sns:
-                log.debug(f"  Device {devices.get(sn).name} ({sn}) excluded")
-                continue
-            if included_sns is not None and sn not in included_sns:
-                continue
-
-            if sn not in matching_sns:
-                matching_sns.append(sn)
-                log.debug(f"  Found matching device: {devices.get(sn).name} ({sn})")
-
-            if not each:
-                return matching_sns
-
-    return matching_sns
